Implement add

+++ com/gitdemo/Calculator.java
@@ -4,10 +4,7 @@ public class Calculator {
     public Calculator() {
     }
 
-    /**
-     * Adds two numbers
-     */
-    public double add(double a, double b) {
+    public int add(int a, int b) {
         return a + b;
     }
 }
